Sean Killeen Resignation from Maintainers Committee
I appreciate the goal and aims of the committee and the work that its members are trying to do. But its current format and impact does not align with the work I want to be doing in this space, and so I fondly and respectfully bid the committee farewell. This absence will be small, because I have been unable to fully participate for some time anyway. I wish you every success in making the .NET Foundation and the .NET ecosystem better for maintainers.